requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

torch==2.6.0+cu124
transformers==4.51.3
accelerate==1.6.0
pandas==2.2.3
tqdm==4.67.1
PyYAML==6.0.2
